+++ core/network/src/main/java/com/school_of_company/network/datasource/expo/ExpoDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.school_of_company.network.datasource.expo
 
 import com.school_of_company.network.api.ExpoAPI
-import com.school_of_company.network.dto.expo.request_response.ExpoRequestAndResponse
+import com.school_of_company.network.dto.expo.all.ExpoRequestAndResponse
 import com.school_of_company.network.dto.expo.response.ExpoIdResponse
 import com.school_of_company.network.dto.expo.response.ExpoListResponse
 import com.school_of_company.network.util.performApiRequest
